Move CanSocket to its own module
