+++ src/socket.rs
@@ -1,0 +1,366 @@
+
+use crate::frame::CanFrame;
+use crate::err::{CanSocketOpenError, ConstructionError};
+use crate::util::{set_socket_option, set_socket_option_mult, system_time_from_timespec};
+use crate::constants::*;
+
+
+use libc::{socket, SOCK_RAW, close, bind, sockaddr, read,
+    write, SOL_SOCKET, SO_RCVTIMEO, timespec, timeval, EINPROGRESS, SO_SNDTIMEO, time_t,
+    suseconds_t, fcntl, F_GETFL, F_SETFL, O_NONBLOCK};
+use nix::net::if_::if_nametoindex;
+use std::{
+    os::raw::{c_int, c_short, c_void, c_uint, c_ulong},
+    fmt,
+    io,
+    time,
+    mem::size_of,
+};
+use std::os::unix::io::{AsRawFd, FromRawFd, IntoRawFd, RawFd};
+
+/// Check an error return value for timeouts.
+///
+/// Due to the fact that timeouts are reported as errors, calling `read_frame`
+/// on a socket with a timeout that does not receive a frame in time will
+/// result in an error being returned. This trait adds a `should_retry` method
+/// to `Error` and `Result` to check for this condition.
+pub trait ShouldRetry {
+    /// Check for timeout
+    ///
+    /// If `true`, the error is probably due to a timeout.
+    fn should_retry(&self) -> bool;
+}
+
+impl ShouldRetry for io::Error {
+    fn should_retry(&self) -> bool {
+        match self.kind() {
+            // EAGAIN, EINPROGRESS and EWOULDBLOCK are the three possible codes
+            // returned when a timeout occurs. the stdlib already maps EAGAIN
+            // and EWOULDBLOCK os WouldBlock
+            io::ErrorKind::WouldBlock => true,
+            // however, EINPROGRESS is also valid
+            io::ErrorKind::Other => {
+                if let Some(i) = self.raw_os_error() {
+                    i == EINPROGRESS
+                } else {
+                    false
+                }
+            }
+            _ => false,
+        }
+    }
+}
+
+impl<E: fmt::Debug> ShouldRetry for io::Result<E> {
+    fn should_retry(&self) -> bool {
+        if let Err(ref e) = *self {
+            e.should_retry()
+        } else {
+            false
+        }
+    }
+}
+
+
+fn c_timeval_new(t: time::Duration) -> timeval {
+    timeval {
+        tv_sec: t.as_secs() as time_t,
+        tv_usec: t.subsec_micros() as suseconds_t,
+    }
+}
+
+#[derive(Debug)]
+#[repr(C)]
+struct CanAddr {
+    _af_can: c_short,
+    if_index: c_int, // address familiy,
+    rx_id: u32,
+    tx_id: u32,
+}
+
+/// A socket for a CAN device.
+///
+/// Will be closed upon deallocation. To close manually, use std::drop::Drop.
+/// Internally this is just a wrapped file-descriptor.
+#[derive(Debug)]
+pub struct CanSocket {
+    fd: c_int,
+}
+
+impl CanSocket {
+    /// Open a named CAN device.
+    ///
+    /// Usually the more common case, opens a socket can device by name, such
+    /// as "vcan0" or "socan0".
+    pub fn open(ifname: &str) -> Result<CanSocket, CanSocketOpenError> {
+        let if_index = if_nametoindex(ifname)?;
+        CanSocket::open_if(if_index)
+    }
+
+    /// Open CAN device by interface number.
+    ///
+    /// Opens a CAN device by kernel interface number.
+    pub fn open_if(if_index: c_uint) -> Result<CanSocket, CanSocketOpenError> {
+        let addr = CanAddr {
+            _af_can: AF_CAN as c_short,
+            if_index: if_index as c_int,
+            rx_id: 0, // ?
+            tx_id: 0, // ?
+        };
+
+        // open socket
+        let sock_fd;
+        unsafe {
+            sock_fd = socket(PF_CAN, SOCK_RAW, CAN_RAW);
+        }
+
+        if sock_fd == -1 {
+            return Err(CanSocketOpenError::from(io::Error::last_os_error()));
+        }
+
+        // bind it
+        let bind_rv;
+        unsafe {
+            let sockaddr_ptr = &addr as *const CanAddr;
+            bind_rv = bind(sock_fd,
+                           sockaddr_ptr as *const sockaddr,
+                           size_of::<CanAddr>() as u32);
+        }
+
+        // FIXME: on fail, close socket (do not leak socketfds)
+        if bind_rv == -1 {
+            let e = io::Error::last_os_error();
+            unsafe {
+                close(sock_fd);
+            }
+            return Err(CanSocketOpenError::from(e));
+        }
+
+        Ok(CanSocket { fd: sock_fd })
+    }
+
+    fn close(&mut self) -> io::Result<()> {
+        unsafe {
+            let rv = close(self.fd);
+            if rv != -1 {
+                return Err(io::Error::last_os_error());
+            }
+        }
+        Ok(())
+    }
+
+    /// Change socket to non-blocking mode
+    pub fn set_nonblocking(&self, nonblocking: bool) -> io::Result<()> {
+        // retrieve current flags
+        let oldfl = unsafe { fcntl(self.fd, F_GETFL) };
+
+        if oldfl == -1 {
+            return Err(io::Error::last_os_error());
+        }
+
+        let newfl = if nonblocking {
+            oldfl | O_NONBLOCK
+        } else {
+            oldfl & !O_NONBLOCK
+        };
+
+        let rv = unsafe { fcntl(self.fd, F_SETFL, newfl) };
+
+        if rv != 0 {
+            return Err(io::Error::last_os_error());
+        }
+        Ok(())
+    }
+
+    /// Sets the read timeout on the socket
+    ///
+    /// For convenience, the result value can be checked using
+    /// `ShouldRetry::should_retry` when a timeout is set.
+    pub fn set_read_timeout(&self, duration: time::Duration) -> io::Result<()> {
+        set_socket_option(self.fd, SOL_SOCKET, SO_RCVTIMEO, &c_timeval_new(duration))
+    }
+
+    /// Sets the write timeout on the socket
+    pub fn set_write_timeout(&self, duration: time::Duration) -> io::Result<()> {
+        set_socket_option(self.fd, SOL_SOCKET, SO_SNDTIMEO, &c_timeval_new(duration))
+    }
+
+    /// Blocking read a single can frame.
+    pub fn read_frame(&self) -> io::Result<CanFrame> {
+        // let mut frame = CanFrame {
+        //     _id: 0,
+        //     _data_len: 0,
+        //     _pad: 0,
+        //     _res0: 0,
+        //     _res1: 0,
+        //     _data: [0; 8],
+        // };
+        let mut frame = CanFrame::default();
+
+        let read_rv = unsafe {
+            let frame_ptr = &mut frame as *mut CanFrame;
+            read(self.fd, frame_ptr as *mut c_void, size_of::<CanFrame>())
+        };
+
+        if read_rv as usize != size_of::<CanFrame>() {
+            return Err(io::Error::last_os_error());
+        }
+
+        Ok(frame)
+    }
+
+    /// Blocking read a single can frame with timestamp
+    ///
+    /// Note that reading a frame and retrieving the timestamp requires two
+    /// consecutive syscalls. To avoid race conditions, exclusive access
+    /// to the socket is enforce through requiring a `mut &self`.
+    pub fn read_frame_with_timestamp(&mut self) -> io::Result<(CanFrame, time::SystemTime)> {
+        let frame = self.read_frame()?;
+
+        let mut ts = timespec { tv_sec: 0, tv_nsec: 0 };
+        let rval = unsafe {
+            libc::ioctl(self.fd, SIOCGSTAMPNS as c_ulong, &mut ts as *mut timespec)
+        };
+
+        if rval == -1 {
+            return Err(io::Error::last_os_error());
+        }
+
+        Ok((frame, system_time_from_timespec(ts)))
+    }
+
+    /// Write a single can frame.
+    ///
+    /// Note that this function can fail with an `EAGAIN` error or similar.
+    /// Use `write_frame_insist` if you need to be sure that the message got
+    /// sent or failed.
+    pub fn write_frame(&self, frame: &CanFrame) -> io::Result<()> {
+        // not a mutable reference needed (see std::net::UdpSocket) for
+        // a comparison
+        // debug!("Sending: {:?}", frame);
+
+        let write_rv = unsafe {
+            let frame_ptr = frame as *const CanFrame;
+            write(self.fd, frame_ptr as *const c_void, size_of::<CanFrame>())
+        };
+
+        if write_rv as usize != size_of::<CanFrame>() {
+            return Err(io::Error::last_os_error());
+        }
+
+        Ok(())
+    }
+
+    /// Blocking write a single can frame, retrying until it gets sent
+    /// successfully.
+    pub fn write_frame_insist(&self, frame: &CanFrame) -> io::Result<()> {
+        loop {
+            match self.write_frame(frame) {
+                Ok(v) => return Ok(v),
+                Err(e) => {
+                    if !e.should_retry() {
+                        return Err(e);
+                    }
+                }
+            }
+        }
+    }
+
+    /// Sets filters on the socket.
+    ///
+    /// CAN packages received by SocketCAN are matched against these filters,
+    /// only matching packets are returned by the interface.
+    ///
+    /// See `CanFilter` for details on how filtering works. By default, all
+    /// single filter matching all incoming frames is installed.
+    pub fn set_filters(&self, filters: &[CanFilter]) -> io::Result<()> {
+        set_socket_option_mult(self.fd, SOL_CAN_RAW, CAN_RAW_FILTER, filters)
+    }
+
+    /// Sets the error mask on the socket.
+    ///
+    /// By default (`ERR_MASK_NONE`) no error conditions are reported as
+    /// special error frames by the socket. Enabling error conditions by
+    /// setting `ERR_MASK_ALL` or another non-empty error mask causes the
+    /// socket to receive notification about the specified conditions.
+    #[inline]
+    pub fn set_error_mask(&self, mask: u32) -> io::Result<()> {
+        set_socket_option(self.fd, SOL_CAN_RAW, CAN_RAW_ERR_FILTER, &mask)
+    }
+
+    /// Enable or disable loopback.
+    ///
+    /// By default, loopback is enabled, causing other applications that open
+    /// the same CAN bus to see frames emitted by different applications on
+    /// the same system.
+    #[inline]
+    pub fn set_loopback(&self, enabled: bool) -> io::Result<()> {
+        let loopback: c_int = if enabled { 1 } else { 0 };
+        set_socket_option(self.fd, SOL_CAN_RAW, CAN_RAW_LOOPBACK, &loopback)
+    }
+
+    /// Enable or disable receiving of own frames.
+    ///
+    /// When loopback is enabled, this settings controls if CAN frames sent
+    /// are received back immediately by sender. Default is off.
+    pub fn set_recv_own_msgs(&self, enabled: bool) -> io::Result<()> {
+        let recv_own_msgs: c_int = if enabled { 1 } else { 0 };
+        set_socket_option(self.fd, SOL_CAN_RAW, CAN_RAW_RECV_OWN_MSGS, &recv_own_msgs)
+    }
+
+    /// Enable or disable join filters.
+    ///
+    /// By default a frame is accepted if it matches any of the filters set
+    /// with `set_filters`. If join filters is enabled, a frame has to match
+    /// _all_ filters to be accepted.
+    pub fn set_join_filters(&self, enabled: bool) -> io::Result<()> {
+        let join_filters: c_int = if enabled { 1 } else { 0 };
+        set_socket_option(self.fd, SOL_CAN_RAW, CAN_RAW_JOIN_FILTERS, &join_filters)
+    }
+}
+
+impl AsRawFd for CanSocket {
+    fn as_raw_fd(&self) -> RawFd {
+        self.fd
+    }
+}
+
+impl FromRawFd for CanSocket {
+    unsafe fn from_raw_fd(fd: RawFd) -> CanSocket {
+        CanSocket { fd, }
+    }
+}
+
+impl IntoRawFd for CanSocket {
+    fn into_raw_fd(self) -> RawFd {
+        self.fd
+    }
+}
+
+impl Drop for CanSocket {
+    fn drop(&mut self) {
+        self.close().ok(); // ignore result
+    }
+}
+
+
+/// CanFilter
+///
+/// Contains an internal id and mask. Packets are considered to be matched by
+/// a filter if `received_id & mask == filter_id & mask` holds true.
+#[derive(Debug, Copy, Clone)]
+#[repr(C)]
+pub struct CanFilter {
+    _id: u32,
+    _mask: u32,
+}
+
+impl CanFilter {
+    /// Construct a new CAN filter.
+    pub fn new(id: u32, mask: u32) -> Result<CanFilter, ConstructionError> {
+        Ok(CanFilter {
+               _id: id,
+               _mask: mask,
+           })
+    }
+}
